Firedrake backend: Move interpolation code
