+++ tests/firedrake/test_base.py
@@ -3,8 +3,7 @@ from tlm_adjoint.firedrake import *
 from tlm_adjoint.firedrake import manager as _manager
 from tlm_adjoint.firedrake.backend import (
     backend_Cofunction, backend_Constant, backend_Function, complex_mode)
-from tlm_adjoint.firedrake.backend_code_generator_interface import (
-    interpolate_expression)
+from tlm_adjoint.firedrake.interpolation import interpolate_expression
 from tlm_adjoint.alias import gc_disabled
 from tlm_adjoint.patch import patch_method
