+++ tlm_adjoint/firedrake/__init__.py
@@ -15,5 +15,6 @@ from .equations import *            # noqa: E402,F401
 from .firedrake_equations import *  # noqa: E402,F401
 from .functions import *            # noqa: E402,F401
 from .hessian_system import *       # noqa: E402,F401
+from .interpolation import *        # noqa: E402,F401
 
 from .backend_patches import *      # noqa: E402,F401
